enviando e-mail após evento de resposta cadastrada

+++ app/Services/ReplySupportService.php
@@ -2,8 +2,12 @@
 namespace App\Services;
 
 use App\DTO\Supports\Replies\CreateReplyDTO;
+use App\Events\SupportReplied;
 use App\Repositories\Contracts\ReplySupportRepository;
+use Illuminate\Support\Facades\Event;
+use Illuminate\Support\Facades\Gate;
 use stdClass;
+use Symfony\Component\Translation\Exception\NotFoundResourceException;
 
 class ReplySupportService {
 
@@ -13,10 +17,19 @@ class ReplySupportService {
     }
 
     public function create(CreateReplyDTO $dto): stdClass {
-        return $this->repository->create($dto);
+        $reply = $this->repository->create($dto);
+        SupportReplied::dispatch($reply->support);
+        return $reply;
     }
 
     public function delete(string $id): bool {
+        $reply = $this->repository->getById($id);
+        if(!$reply) {
+            throw new NotFoundResourceException('Resposta não encontrada');
+        }
+        if(Gate::denies('owner', $reply->user['id'])) {
+            abort(403, 'Não Autorizado');
+        }
         return $this->repository->delete($id);
     }
 }
